requirements.txt
flet==0.22.*
types-python-dateutil==2.9.0.20240316
python-dateutil==2.9.0.post0